added gps version

+++ CANdaemon/canable.sh
@@ -3,10 +3,18 @@
 user=`id -u`
 
 if [ $user -ne 0 ]
-  then echo "Run as root"
-  exit
+    then 
+        echo "Run as root"
+        exit
+fi
+
+if [ $# -eq 0 ]
+    then 
+        echo "No argument supplied"
+        exit
 fi
 
 killall -q slcand
-slcand -o -c -f -s6 /dev/serial/by-id/usb-CANtact_CANtact_dev_00000000001A-if00 can0
+slcand -o -c -f -s8 $1 can0
+sleep 1
 ip link set can0 up
